minor changes in reown_cli
