+++ packages/reown_cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+- Minor fixes
+
 ## 0.0.2
 
 - Minor fixes
